Improved the way the dev session works
Now if DEBUG the app logs into a different facebook application
instance - so it can use a local url.

Flask sets the local url, and the facebook object is initated with
O-auth for that url, and app combination.

This makes testing much more realistic (because you are actually
testing on a real user).